backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
python-dotenv==1.0.1
huggingface_hub==0.30.2
pydantic[email]==2.10.4